Move classes to catalog namespace
