+++ tests/back/Acceptance/Catalog/Category/InMemoryCategoryRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Test\Acceptance\Category;
+namespace Akeneo\Test\Acceptance\Catalog\Category;
 
 use Akeneo\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Akeneo\Component\StorageUtils\Saver\SaverInterface;
