+++ tests/back/Acceptance/Catalog/Product/InMemoryProductRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Test\Acceptance\Product;
+namespace Akeneo\Test\Acceptance\Catalog\Product;
 
 use Akeneo\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Akeneo\Component\StorageUtils\Saver\SaverInterface;
