+++ tests/back/Acceptance/Catalog/Product/Model/InMemoryProductModelRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Test\Acceptance\ProductModel;
+namespace Akeneo\Test\Acceptance\Catalog\Product\Model;
 
 use Akeneo\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Akeneo\Component\StorageUtils\Saver\SaverInterface;
