+++ tests/back/Acceptance/spec/Catalog/Group/Type/InMemoryGroupTypeRepositorySpec.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace spec\Akeneo\Test\Acceptance\Catalog;
+namespace spec\Akeneo\Test\Acceptance\Catalog\Group\Type;
 
 use Akeneo\Component\StorageUtils\Saver\SaverInterface;
-use Akeneo\Test\Acceptance\Catalog\InMemoryGroupTypeRepository;
+use Akeneo\Test\Acceptance\Catalog\Group\Type\InMemoryGroupTypeRepository;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Entity\GroupType;
 use Pim\Component\Catalog\Repository\GroupTypeRepositoryInterface;
