+++ tests/back/Acceptance/spec/Catalog/Product/Model/InMemoryProductModelRepositorySpec.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace spec\Akeneo\Test\Acceptance\ProductModel;
+namespace spec\Akeneo\Test\Acceptance\Catalog\Product\Model;
 
 use Akeneo\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Akeneo\Component\StorageUtils\Saver\SaverInterface;
 use Akeneo\Test\Acceptance\Common\NotImplementedException;
-use Akeneo\Test\Acceptance\ProductModel\InMemoryProductModelRepository;
+use Akeneo\Test\Acceptance\Catalog\Product\Model\InMemoryProductModelRepository;
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Model\FamilyVariant;
 use Pim\Component\Catalog\Model\ProductModel;
