+++ tests/back/Acceptance/spec/Structure/Attribute/Group/InMemoryAttributeGroupRepositorySpec.php
