+++ tests/back/Acceptance/spec/Structure/Attribute/InMemoryAttributeRepositorySpec.php
